researchpilot/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
python-multipart==0.0.22
sentence-transformers==2.2.2
chromadb==0.4.24
pypdf==3.17.1
pydantic==2.5.0
pydantic-settings==2.1.0
numpy==1.24.3
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pyppeteer==0.2.6
python-docx==0.8.11
beautifulsoup4==4.12.2
# Optional fast in-process PDF renderer. Requires system libraries (cairo, pango).
weasyprint==57.1
# sqlalchemy version is managed by sqlmodel compatibility (<=1.4.x)
groq>=1.0.0
